requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
certifi==2025.1.31
distro==1.9.0
Django==5.1.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
Markdown==3.7
nest-asyncio==1.6.0
openai==1.61.0
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==21.10
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
redis==5.0.1
dj-database-url==2.1.0
django-redis==5.4.0
